Handling window-aux.js changes in the remote transition while calendar is not yet changed
