+++ js/classes/FlexibleDayCalendar.js
@@ -9,8 +9,28 @@ import {
 } from '../time-math.js';
 import { getDateStr, getMonthLength } from '../date-aux.js';
 import { generateKey } from '../date-db-formatter.js';
-import { showDialog } from '../window-aux.js';
 import { BaseCalendar } from './BaseCalendar.js';
+
+/// Compatiblity block - to be removed in the migration of calendar to non-remote electron
+const { remote } = require('electron');
+const { BrowserWindow, dialog } = remote;
+
+/**
+ * Opens an electron dialog, based on the options, and performs the successCallback after promise is resolved.
+ * @param {Object.<string, any>} options
+ * @param {function} successCallback
+ */
+ function showDialog(options, successCallback)
+ {
+     options['title'] = options['title'] || 'Time to Leave';
+     dialog.showMessageBox(BrowserWindow.getFocusedWindow(), options)
+         .then(successCallback)
+         .catch(err =>
+         {
+             console.log(err);
+         });
+ }
+////
 
 class FlexibleDayCalendar extends BaseCalendar
 {
