+++ js/window-aux.js
@@ -5,22 +5,6 @@ const BrowserWindow = (electron || electron.remote).BrowserWindow;
 const dialog = (electron || electron.remote).dialog;
 
 /**
- * Opens an electron dialog, based on the options, and performs the successCallback after promise is resolved.
- * @param {Object.<string, any>} options
- * @param {function} successCallback
- */
-function showDialog(options, successCallback)
-{
-    options['title'] = options['title'] || 'Time to Leave';
-    dialog.showMessageBox(BrowserWindow.getFocusedWindow(), options)
-        .then(successCallback)
-        .catch(err =>
-        {
-            console.log(err);
-        });
-}
-
-/**
  * Opens an electron dialog, based on the options, and returns the promise.
  * @param {Object.<string, any>} options
  * @return {Promise}
@@ -30,7 +14,6 @@ function showDialogSync(options)
     options['title'] = options['title'] || 'Time to Leave';
     return dialog.showMessageBox(BrowserWindow.getFocusedWindow(), options)
 }
-
 
 /**
  * Opens an electron dialog just like a JS alert().
@@ -47,6 +30,5 @@ function showAlert(message)
 
 export {
     showAlert,
-    showDialog,
     showDialogSync
 };
